# docs/specs/swagger.yaml
{"openapi":"3.0.1","info":{"title":"finalchallengedemo","description":"finalchallengedemo","contact":{"name":"finalchallengedemo","email":"romy.molano@mercadolibre.com.co"},"version":"last"},"servers":[{"url":"https://null_finalchallengedemo.furyapps.io","description":"Scope null"}],"paths":{"/api/v1/parts/update":{"put":{"tags":["part-controller"],"operationId":"update","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PartDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/parts/save":{"post":{"tags":["part-controller"],"operationId":"savePart","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PartDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"operationId":"ping","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/parts/list":{"get":{"tags":["part-controller"],"operationId":"getParts","parameters":[{"name":"filter","in":"query","required":true,"schema":{"$ref":"#/components/schemas/PartFilterDTO"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/parts/find/{id}":{"get":{"tags":["part-controller"],"operationId":"findPart","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/parts/delete/{id}":{"delete":{"tags":["part-controller"],"operationId":"deleteStudent","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"PartDTO":{"type":"object","properties":{"partCode":{"type":"integer","format":"int64"},"description":{"type":"string"},"widthDimension":{"type":"integer","format":"int32"},"tallDimension":{"type":"integer","format":"int32"},"longDimension":{"type":"integer","format":"int32"},"netWeight":{"type":"integer","format":"int32"}}},"PartFilterDTO":{"type":"object","properties":{"queryType":{"type":"string"},"date":{"type":"string","format":"date"},"order":{"type":"integer","format":"int32"}}}}}}